chore: move semver out of @kumahq/config
