+++ packages/kuma-gui/src/app/kuma/index.ts
@@ -8,6 +8,7 @@ import { ApiError } from '@/app/kuma/services/kuma-api/ApiError'
 import KumaApi from '@/app/kuma/services/kuma-api/KumaApi'
 import { RestClient } from '@/app/kuma/services/kuma-api/RestClient'
 import type { ServiceDefinition } from '@kumahq/container'
+export * from './utils'
 
 type Token = ReturnType<typeof token>
